+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## WIP
+- Add filter `Duplicate`.
+
 ## v1.8.2 (2020-04-03)
 - Fix a bug in the filename filter config parsing algorithm with digits-only filenames.